+++ src/main/java/ru/forum/whale/space/api/repository/ChatRepository.java
@@ -9,12 +9,12 @@ import ru.forum.whale.space.api.model.User;
 import java.util.List;
 import java.util.Optional;
 
-public interface ChatRepository extends JpaRepository<Chat, Integer> {
+public interface ChatRepository extends JpaRepository<Chat, Long> {
     Optional<Chat> findByUser1AndUser2(User user1, User user2);
 
     @Query("From Chat c LEFT JOIN FETCH c.messages WHERE c.id = :chatId")
-    Optional<Chat> findByIdWithMessages(@Param("chatId") Integer chatId);
+    Optional<Chat> findByIdWithMessages(@Param("chatId") Long chatId);
 
     @Query("FROM Chat c LEFT JOIN FETCH c.messages WHERE c.user1.id = :userId OR c.user2.id = :userId")
-    List<Chat> findAllByUserIdWithMessages(@Param("userId") Integer userId);
+    List<Chat> findAllByUserIdWithMessages(@Param("userId") Long userId);
 }
